libs/AI Agent In Practice/第三章 Function Calling MCP 与工具使用/3.3 _动手_给 LLM 接入搜索 - 计算器 - 数据库工具/requirements.txt
python-dotenv>=1.0.0
litellm>=1.40.0
tavily-python>=0.3.0
e2b-code-interpreter>=0.5.0
sqlalchemy>=2.0.0
openai>=1.0.0
pytest>=7.0.0
